Brought code into compliance with policies enforced by swiftlint, an aspect of which required breaking test cases into smaller test cases.
